+++ config/laravel-svelte.php
@@ -9,23 +9,23 @@ return [
         // We install Svelte and Typescript packages regardless of the options chosen.
         'base' => [
             '@inertiajs/svelte' => '^1.2.0',
-            '@sveltejs/vite-plugin-svelte' => '^3.1.1',
+            '@sveltejs/vite-plugin-svelte' => '^3.1.2',
             '@tsconfig/svelte' => '^5.0.4',
-            'svelte' => '^4.2.18',
-            'svelte-check' => '^3.8.4',
+            'svelte' => '^4.2.19',
+            'svelte-check' => '^4.0.1',
             'svelte-preprocess' => '^6.0.2',
-            'typescript' => '^5.5.3',
+            'typescript' => '^5.5.4',
         ],
 
         'eslint' => [
-            '@eslint/js' => '^9.7.0',
+            '@eslint/js' => '^9.10.0',
             '@types/eslint__js' => '^8.42.3',
-            'eslint' => '^8.57.0',
+            'eslint' => '^9.10.0',
             'eslint-plugin-svelte' => '^2.43.0',
-            'typescript-eslint' => '^7.16.1',
+            'typescript-eslint' => '^8.4.0',
 
             'cypress' => [
-                'eslint-plugin-cypress' => '^3.3.0',
+                'eslint-plugin-cypress' => '^3.5.0',
             ],
 
             'prettier' => [
@@ -37,12 +37,12 @@ return [
         'prettier' => [
             'prettier' => '^3.3.3',
             'prettier-plugin-svelte' => '^3.2.6',
-            'prettier-plugin-tailwindcss' => '^0.6.5',
+            'prettier-plugin-tailwindcss' => '^0.6.6',
         ],
 
         'cypress' => [
             '@faker-js/faker' => '^8.4.1',
-            'cypress' => '^13.13.1',
+            'cypress' => '^13.14.2',
         ],
 
         'playwright' => [